requirements.txt
aiohttp==3.7.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dill==0.3.2
distro==1.5.0
flake8==3.8.4
idna==2.10
Jinja2==2.11.2
junit-xml==1.9
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.0.0
mypy-extensions==0.4.3
pathspec==0.8.0
psutil==5.7.3
pyats==20.10
pyats.aereport==20.10
pyats.aetest==20.10
pyats.async==20.10
pyats.connections==20.10
pyats.datastructures==20.10
pyats.easypy==20.10
pyats.kleenex==20.10
pyats.log==20.10
pyats.reporter==20.10
pyats.results==20.10
pyats.tcl==20.10
pyats.topology==20.10
pyats.utils==20.10
pycodestyle==2.6.0
pyflakes==2.2.0
python-engineio==3.13.2
python-socketio==4.6.0
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
unicon==20.10
unicon.plugins==20.10
urllib3==1.25.11
virl2-client==2.1.0
yamllint==1.25.0
yarl==1.6.2
